requirements.txt
matplotlib
python-can
python-can[serial]
